pyproject.toml
[project]
name = "velide-middleware"
version = "3.2.0"
description = "Middleware do Velide para realizar integrações com ERPs e sistemas locais."
readme = "README.md"
# 1. PYTHON 3.8 COMPATIBILITY
# This tells uv/pip to ONLY install packages compatible with 3.8
requires-python = ">=3.8"

# 2. AUTOMATIC DEPENDENCY STORAGE
# You do NOT edit this manually. Run `uv add package_name` 
# and it will populate this list with 3.8-compatible versions.
dependencies = [
    # Example: "requests>=2.31.0",
    "annotated-types>=0.7.0",
    "anyio>=4.5.2",
    "cabarchive>=0.2.4",
    "certifi>=2026.1.4",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.4",
    "colorama>=0.4.6",
    "cryptography>=46.0.3",
    "cx-freeze>=6.15.12",
    "cx-logging>=3.2.0",
    "fdb>=2.0.4",
    "filelock>=3.16.1",
    "gql[websockets]>=3.5.3",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "idna>=3.11",
    "lief>=0.17.1",
    "packaging>=25.0",
    "pillow>=10.4.0",
    "pycparser>=2.23",
    "pydantic>=2.10.6",
    "pydantic-core>=2.27.2",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.9.0",
    "pyqt5>=5.15.11",
    "pyqt5-qt5==5.15.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.3",
    "qrcode>=7.4.2",
    "requests>=2.32.4",
    "screeninfo>=0.8.1",
    "setuptools>=69.5.1",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.45",
    "sqlalchemy-firebird>=2.1",
    "striprtf>=0.0.29",
    "tdqm>=0.0.1",
    "tqdm>=4.67.1",
    "urllib3>=2.2.3",
    "watchdog>=4.0.2",
]

[project.optional-dependencies]
# Dependencies just for development (testing, linting)
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

# --- TOOL CONFIGURATION SECTIONS ---

[tool.pytest.ini_options]
# 3. REPLACES pytest.ini
# This points pytest to your source and test folders
pythonpath = "src"
testpaths = ["tests"]

[tool.mypy]
# 1. THE ENFORCER: Fail if code isn't 3.8 compatible
python_version = "3.8"

# 2. THE MAP: Tell Mypy where your code lives
mypy_path = "src"
explicit_package_bases = true
namespace_packages = true

# 3. THE NOISE FILTER: Stop complaining about missing library stubs
ignore_missing_imports = true

# 4. STRICTNESS (Optional but recommended):
# Prevent you from returning None when you said you'd return a string, etc.
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
# 5. RUFF CONFIGURATION
# Targeting Python 3.8 for linting rules
target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (F) and pycodestyle (E) codes by default
select = ["E", "F"]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pyqt5-stubs>=5.15.6.0",
    "ruff>=0.14.10",
    "types-pyyaml>=6.0.12.20241230",
    "types-requests>=2.32.0.20241016",
]
